brcm: rename firmware names expressing host interface type
The firmware name used in brcmfmac for SDIO devices was named
generically as brcmfmac-sdio.bin. To use the brcmfmac SDIO driver
the user had to copy the appropriate firmware to brcmfmac-sdio.bin.

This turned out to be inconvenient for developers switching different
devices on their system so brcmfmac was changed to use chip specific
filenames according format brcmfmac<chipid>-sdio.bin. This patch renames
the existing SDIO firmware files to match this format.

Signed-off-by: Arend van Spriel <arend@broadcom.com>
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>