s5p-mfc: Add s5p-mfc-v6-v2 firmware
This is the newer version of mfc-v6 firmware which can
be used in exynos5250 based boards. This has many bug fixes
and enhancements. To maintain backward compatibility, the older
firmware file is still retained and the new file is added with
the name s5p-mfc-v6-v2.fw. This new firmware handling is supported
with the below patches in s5p-mfc driver:
s5p-mfc: Support multiple firmware sub-versions
s5p-mfc: Add init buffer cmd to MFCV6

Signed-off-by: Arun Kumar K <arun.kk@samsung.com>
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>