linux-firmware: update audio firmware for Braswell platform
This updates the firmware for BSW platform. This update contains fixes
for some randome noise coming during suspend/resume and headset plug
events.

The firmware version is also added to WHENCE file. The version is
v01.0B.02.02.

The md5sum of the file is 7344444bff5257718bd9e56d961e899c

Signed-off-by: Vinod Koul <vinod.koul@intel.com>