brcmfmac: firmware refresh for BCM43602 PCIE devices
A number of end-users reported issues with BCM43602 on the 2015
MacBook Pro. It turned out the previously released firmware did
not properly support that device [1]. Hence an update of the STA
firmware and while at it refresh the AP firmware as well. The
latter has been tested using Netgear R8000. The STA firmware has
been tested on MacBook Pro.

[1] https://bugzilla.kernel.org/show_bug.cgi?id=100201

Cc: Rafal Milecki <zajec5@gmail.com>
Tested-by: Knuth Posern <knuth@posern.org>
Reviewed-by: Pieter-Paul Giesberts <pieterpg@broadcom.com>
Signed-off-by: Arend van Spriel <arend@broadcom.com>
Signed-off-by: Kyle McMartin <kyle@kernel.org>