usbdux: usbduxsigma: created matching binary of the firmware
The firmware source (alongside with the usbduxsigma.c)
was updated in 2015 during a major rewrite but
its binary was still the old one which caused the
kernel driver to fail. This fix provides the matching
firmware binary. This was created by invoking
"make -f Makefile_dux" which creates the firmware binary.

Signed-off-by: tech@linux-usb-daq.co.uk
Signed-off-by: Kyle McMartin <kyle@kernel.org>