linux-firmware/i915: Add Geminilake DMC version 1.04
DMC provides additional graphics low-power idle states. It provides
capability to save and restore display registers across these low-power
states independently from the OS/Kernel

This is the first release of DMC for Geminilake.

This major version will be just upgraded on code when it is required
software changes for API/ABI compatibility.

Cc: Rodrigo Vivi <rodrigo.vivi@intel.com>
Signed-off-by: Anusha Srivatsa <anusha.srivatsa@intel.com>