linux-firmware: Add AMD SEV firmware
Add AMD Secure Encrypted Virtualization (SEV) firmware to allow for the
updating of the SEV firmware outside of BIOS releases. Naming convention
and directory structure follows the convention defined in the kernel
driver in drivers/crypto/ccp/psp-dev.c

SEV firmware files are to be placed in the "amd" directory and names
should follow the pattern: amd_sev_fam<FAMILY>h_model<MODEL>h.sbin

where FAMILY is the two-digit hexadecimal value of the target processor
family (e.g. 17)

where MODEL is the two-digit hexadecimal value of the target processor
model (e.g. 01) or where the least significant nibble of the two-digit
hexadecimal value can be an 'x' to indicate a range of models
(e.g. 0x for 00 to of).

The firmware provided is version 0.17 build 11 and is for AMD family
17h processors with models in the range 00h to 0fh.

Signed-off-by: John Allen <john.allen@amd.com>
Signed-off-by: Josh Boyer <jwboyer@kernel.org>