amlogic: add video decoder firmwares
Add the firmwares for various video compression standards that are
required to decode video on Amlogic SoCs using hardware acceleration.

They are listed under a custom redistributable license provided by
Amlogic.

Although the filelist is not complete yet (there are a _lot_ of
different codec/SoC combinations), the firmwares added here are
supported and tested with the work-in-progress mainline driver.

The latest submission for the mainline driver can be found here[1], and
the development tree can be found here[2].

[1] https://patchwork.kernel.org/cover/10879959/
[2] https://github.com/Elyotna/linux

Cc: Arden Jin <Arden.Jin@amlogic.com>
Signed-off-by: Maxime Jourdan <mjourdan@baylibre.com>